Refactor package message to have a more functional interface
MessagePipe has been moved to package message and has been completely
rewritten.
- Viewers are now referred to as receivers.
- Now all you need to make a MessagePipe is a stream of receivers.
- You can no longer explicitly add or remove receivers from a
	MessagePipe. Instead you must define a better MessagePipe.
- You can now perform set operations such as union or intersection
	on MessagePipes.
- ChatMessages are now supplied to a MessagePipe with a
	MessageSupplier.

The goal of these changes is to create lazy-evaluated message
formatting. Operations that may be expensive like parsing placeholders
should be deferred until we are absolutely sure that the receiver
should be receiving the message.

Furthermore, these changes allow quick creation of new channel types.
For example, if a user wants to add a chat channel that can only be
viewed by members of your faction that are in the same world as you,
they should be able to simply intersect a faction pipe with a world
pipe.